pyproject.toml
[tool.poetry]
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
openai = "^1.82.0"
streamlit = "^1.45.1"
python-dotenv = "^1.1.0"
groq = "^0.28.0"
docling = "^2.37.0"
chromadb = "^1.0.13"
chromadbx = "^0.1.0"
elasticsearch = "8.15.0"
sentence-transformers = "^5.0.0"

[tool.poetry.extras]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"